+++ src/cli/command-registrar-builder.js
@@ -33,6 +33,7 @@ import Bind from './commands/public-cmds/bind-cmd';
 import Watch from './commands/public-cmds/watch-cmd';
 import Add from './commands/public-cmds/add-cmd';
 import Untrack from './commands/public-cmds/untrack-cmd';
+import Move from './commands/public-cmds/move-cmd';
 
 export default function registerCommands(): CommandRegistrar {
   return new CommandRegistrar(BIT_USAGE, BIT_DESCRIPTION, BIT_VERSION, [
@@ -67,6 +68,7 @@ export default function registerCommands(): CommandRegistrar {
     new Bind(),
     new Watch(),
     new Add(),
-    new Untrack()
+    new Untrack(),
+    new Move()
   ]);
 }