+++ src/consumer/bit-map/bit-map.js
@@ -5,10 +5,11 @@ import find from 'lodash.find';
 import pickBy from 'lodash.pickby';
 import json from 'comment-json';
 import logger from '../../logger/logger';
-import { BIT_MAP, DEFAULT_INDEX_NAME, BIT_JSON, COMPONENT_ORIGINS, DEPENDENCIES_DIR, AUTO_GENERATED_MSG, DEFAULT_SEPARATOR, DEFAULT_INDEX_EXTS } from '../../constants';
+import { BIT_MAP, DEFAULT_INDEX_NAME, BIT_JSON, COMPONENT_ORIGINS, DEPENDENCIES_DIR, AUTO_GENERATED_MSG,
+  DEFAULT_SEPARATOR, DEFAULT_INDEX_EXTS } from '../../constants';
 import { InvalidBitMap, MissingMainFile, MissingBitMapComponent } from './exceptions';
 import { BitId } from '../../bit-id';
-import { readFile, outputFile, pathNormalizeToLinux, pathJoinLinux } from '../../utils';
+import { readFile, outputFile, pathNormalizeToLinux, pathJoinLinux, isDir } from '../../utils';
 
 const SHOULD_THROW = true;
 
@@ -162,16 +163,24 @@ export default class BitMap {
     });
   }
 
+  _getPathWithoutRootDir(rootDir, filePath) {
+    const newPath = path.relative(rootDir, filePath);
+    if (newPath.startsWith('..')) {
+      // this is forbidden for security reasons. Allowing files to be written outside the components directory may
+      // result in overriding OS files.
+      throw new Error(`unable to add file ${filePath} because it's located outside the component root dir ${rootDir}`);
+    }
+    return newPath;
+  }
+
   _changeFilesPathAccordingToItsRootDir(existingRootDir, files) {
+    const changes = [];
     files.forEach((file) => {
-      const newRelativePath = path.relative(existingRootDir, file.relativePath);
-      if (newRelativePath.startsWith('..')) {
-        // this is forbidden for security reasons. Allowing files to be written outside the components directory may
-        // result in overriding OS files.
-        throw new Error(`unable to add file ${file.relativePath} because it's located outside the component root dir ${existingRootDir}`);
-      }
-      file.relativePath = newRelativePath;
+      const newPath = this._getPathWithoutRootDir(existingRootDir, file.relativePath);
+      changes.push({ from: file.relativePath, to: newPath });
+      file.relativePath = newPath;
     });
+    return changes;
   }
 
   /**
@@ -393,5 +402,55 @@ export default class BitMap {
     logger.debug('writing to bit.map');
     this.modifyComponentsToLinuxPath(this.components);
     return outputFile(this.mapPath, AUTO_GENERATED_MSG + JSON.stringify(this.components, null, 4));
+  }
+
+  _findFileOfComponentMap(componentMap: ComponentMap, fileName: string): ComponentMapFile {
+    return componentMap.files.find((file) => {
+      const filePath = componentMap.rootDir ? path.join(componentMap.rootDir, file.relativePath) : file.relativePath;
+      return filePath === fileName;
+    });
+  }
+
+  _updateFileLocation(componentMap, fileFrom, fileTo) {
+    const currentFile = this._findFileOfComponentMap(componentMap, fileFrom);
+    if (!currentFile) throw new Error(`the file ${fileFrom} is untracked`);
+    const rootDir = componentMap.rootDir;
+    const newLocation = rootDir ? this._getPathWithoutRootDir(rootDir, fileTo) : fileTo;
+    logger.debug(`updating file location from ${currentFile.relativePath} to ${newLocation}`);
+    if (componentMap.mainFile === currentFile.relativePath) componentMap.mainFile = newLocation;
+    const changes = [{ from: currentFile.relativePath, to: newLocation }];
+    currentFile.relativePath = newLocation;
+    return changes;
+  }
+
+  _updateDirLocation(componentMap, dirFrom, dirTo) {
+    const changes = [];
+    const rootDir = componentMap.rootDir;
+    if (rootDir && rootDir === dirFrom) {
+      componentMap.rootDir = dirTo;
+      return this._changeFilesPathAccordingToItsRootDir(dirTo, componentMap.files);
+    }
+    let areFilesChanged = false;
+    componentMap.files.forEach((file) => {
+      const filePath = rootDir ? path.join(rootDir, file.relativePath) : file.relativePath;
+      if (filePath.startsWith(dirFrom)) {
+        areFilesChanged = true;
+        const fileTo = filePath.replace(dirFrom, dirTo);
+        const newLocation = rootDir ? this._getPathWithoutRootDir(rootDir, fileTo) : fileTo;
+        logger.debug(`updating file location from ${file.relativePath} to ${newLocation}`);
+        if (componentMap.mainFile === file.relativePath) componentMap.mainFile = newLocation;
+        changes.push({ from: file.relativePath, to: newLocation });
+        file.relativePath = newLocation;
+      }
+    });
+    if (!areFilesChanged) {
+      throw new Error(`neither one of the files use the directory ${dirFrom}`);
+    }
+    return changes;
+  }
+
+  updatePathLocation(componentMap, from, to) {
+    if (isDir(from)) return this._updateDirLocation(componentMap, from, to);
+    return this._updateFileLocation(componentMap, from, to);
   }
 }